add basic pieces
