--- Project/Assets/Scripts/Chick.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c0870161d900f374783202f00c6cbf42
