--- Project/Assets/Scripts/IPiece.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d5c77410ce4b1094a9f1047763021bfd
