+++ Project/Assets/Scripts/PIeces/King.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 358d74ed1b0677c45aa0257ad8fcf377
