+++ Project/Assets/Scripts/PIeces/Queen.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 99df35a25ef4ac944bc8f96d1c2b72ee
