+++ Project/Assets/Scripts/PIeces/Soldier.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1cd97fd2b9776b14587f3a93ffbea9fd
