+++ Project/Library/Bee/1900b0aE.dag.json
@@ -4923,7 +4923,7 @@
       "DisplayName": "Writing Assembly-CSharp.rsp",
       "ActionType": "WriteFile",
       "PayloadOffset": 98143,
-      "PayloadLength": 34936,
+      "PayloadLength": 35052,
       "PayloadDebugContentSnippet": "-target:library\n-out:\"Library/",
       "Inputs": [],
       "InputFlags": [],
@@ -4939,7 +4939,7 @@
       "Annotation": "WriteText Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp2",
       "DisplayName": "Writing Assembly-CSharp.rsp2",
       "ActionType": "WriteFile",
-      "PayloadOffset": 133166,
+      "PayloadOffset": 133282,
       "PayloadLength": 0,
       "PayloadDebugContentSnippet": "",
       "Inputs": [],
@@ -5199,15 +5199,21 @@
         "Library/Bee/artifacts/1900b0aE.dag/Unity.Multiplayer.Center.Editor.ref.dll",
         "Assets/Scripts/Board.cs",
         "Assets/Scripts/Cell.cs",
-        "Assets/Scripts/Chick.cs",
-        "Assets/Scripts/IPiece.cs",
         "Assets/Scripts/PieceMover.cs",
+        "Assets/Scripts/PIeces/General.cs",
+        "Assets/Scripts/PIeces/IPiece.cs",
+        "Assets/Scripts/PIeces/King.cs",
+        "Assets/Scripts/PIeces/Queen.cs",
+        "Assets/Scripts/PIeces/Soldier.cs",
         "C:/Program Files/Unity/Hub/Editor/6000.1.15f1/Editor/Data/Tools/Unity.SourceGenerators/Unity.Properties.SourceGenerator.dll",
         "C:/Program Files/Unity/Hub/Editor/6000.1.15f1/Editor/Data/Tools/Unity.SourceGenerators/Unity.SourceGenerators.dll",
         "C:/Program Files/Unity/Hub/Editor/6000.1.15f1/Editor/Data/Tools/Unity.SourceGenerators/Unity.UIToolkit.SourceGenerator.dll",
         "Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp"
       ],
       "InputFlags": [
+        0,
+        0,
+        0,
         0,
         0,
         0,
@@ -5517,7 +5523,7 @@
       "Annotation": "WriteText Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.dll.mvfrm.rsp",
       "DisplayName": "Writing Assembly-CSharp.dll.mvfrm.rsp",
       "ActionType": "WriteFile",
-      "PayloadOffset": 133262,
+      "PayloadOffset": 133378,
       "PayloadLength": 10240,
       "PayloadDebugContentSnippet": "Library\\Bee\\artifacts\\mvdfrm\\U",
       "Inputs": [],
