+++ Project/Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp
@@ -374,9 +374,12 @@
 -analyzer:"C:/Program Files/Unity/Hub/Editor/6000.1.15f1/Editor/Data/Tools/Unity.SourceGenerators/Unity.UIToolkit.SourceGenerator.dll"
 "Assets/Scripts/Board.cs"
 "Assets/Scripts/Cell.cs"
-"Assets/Scripts/Chick.cs"
-"Assets/Scripts/IPiece.cs"
 "Assets/Scripts/PieceMover.cs"
+"Assets/Scripts/PIeces/General.cs"
+"Assets/Scripts/PIeces/IPiece.cs"
+"Assets/Scripts/PIeces/King.cs"
+"Assets/Scripts/PIeces/Queen.cs"
+"Assets/Scripts/PIeces/Soldier.cs"
 -langversion:9.0
 /deterministic
 /optimize+
